+++ src/components/Accordion/AccordionContent.tsx
@@ -2,16 +2,17 @@
 
 import React from "react";
 import Expandable from "components/_private/Expandable";
+import View from "components/View";
 import AccordionContext from "./Accordion.context";
 import type * as T from "./Accordion.types";
 
 const AccordionContent: React.FC<T.ContentProps> = (props) => {
 	const { children } = props;
-	const { active, triggerId, contentId } = React.useContext(AccordionContext);
+	const { active, triggerId, contentId, gap } = React.useContext(AccordionContext);
 
 	return (
 		<Expandable active={active} attributes={{ "aria-labelledby": triggerId, id: contentId }}>
-			{children}
+			{gap ? <View paddingTop={gap}>{children}</View> : children}
 		</Expandable>
 	);
 };